refac: move core to estring crate
