+++ src/estr.rs
@@ -1,0 +1,6 @@
+#[cfg(feature = "vec")]
+pub mod vec;
+#[cfg(feature = "vec")]
+pub use vec::{CommaVec, SemiVec};
+
+pub use estring::core::*;
